requirements.txt
ccxt
pyyaml
pushbullet.py
peewee